portfolio_api/requirements.txt
psycopg2-binary
fastapi
uvicorn[standard]
